### 2019-03-14 #1318 @whedon
[PRE REVIEW]: RAFF: Robust Algebraic Fitting Function in Julia
**Submitting author:** @fsobral (<a href="http://orcid.org/0000-0003-4963-0946">Francisco Nogueira Calmon Sobral</a>)
**Repository:** <a href="https://github.com/fsobral/RAFF.jl" target ="_blank">https://github.com/fsobral/RAFF.jl</a>
**Version:** v0.4.1
