### 2019-03-17 #1318 @fsobral
Thanks, this is excellent news. If you need a suggestion for a referee, I have already talked to @abelsiqueira, who was interesting in reviewing the algorithm.
