### 2019-03-27 #1318 @poulson
Thank you for your responses, Scott (@stsievert), Jakob (@jakobbossek), and Jon (@malmaud). Jakob, I will try to move forward with finding other reviewers and get back to you if for some reason they cannot be found in the next two weeks.
