### 2019-03-27 #1318 @poulson
Hi Christopher (@ChrisRackauckas) -- you are one of the few reviewers listing both Julia and optimization in your expertise. Would you be able to help with this submission?

Hi Rowan (@rowanc1) -- you have listed both optimization and Julia in your expertise and I was hoping that you would be interesting and able to serve as a reviewer for this submission. Would you be willing to review this submission?

Hi Matheiu (@matbesancon -- I apologize for repeatedly requesting you to review, but you are one of the few reviewers explicitly listing both Julia and optimization in your expertise. Would you be willing and able to review this submission?

Hi Oscar (@oesteban) -- would you be interested and able to review this submission? I believe that your expertise in optimization would be of help, even though you list Python and not Julia in your core language expertise.

Hi Andreas (@andim) -- would you be interested and able to review this submission? With your optimization, python, and C expertise, I believe you could be of help here.
