### 2019-04-15 fsobral/RAFF.jl#8 @matbesancon
[JOSS] Project.toml version mismatch
This issue is raised as part of the review process for https://github.com/openjournals/joss-reviews/issues/1385

The Project.toml file still has 0.4.0 as version number, it should be changed to 0.4.1 which corresponds to the latest release
