### 2019-04-15 fsobral/RAFF.jl#8 @fsobral
Thank you. I have updated both Project and the tag names to v0.4.2.
