### 2019-04-15 fsobral/RAFF.jl#8 @matbesancon
Great!
