### 2019-04-15 89480d6
Change function names to follow Julia style guide. #7.
