+++ src/generator.jl
@@ -1,4 +1,4 @@
-export generateTestProblems, generateNoisyData
+export generate_test_problems, generate_noisy_data
 
 # This dictionary represents the list of models used in the tests
 # Return the tuple (n, model, model_str)
@@ -18,7 +18,7 @@ const model_list = Dict(
 
 """
 
-    generateTestProblems(datFilename::String, solFilename::String,
+    generate_test_problems(datFilename::String, solFilename::String,
                          model::Function, modelStr::String, n::Int,
                          np::Int, p::Int)
 
@@ -40,7 +40,7 @@ Generate random data files for testing fitting problems.
     approach.
 
 """
-function generateTestProblems(datFilename::String,
+function generate_test_problems(datFilename::String,
                               solFilename::String, model::Function,
                               modelStr::String,
                               n::Int, np::Int, p::Int;
@@ -63,7 +63,7 @@ function generateTestProblems(datFilename::String,
     
     open(datFilename, "w") do data
     
-        vdata, xsol, outliers = generateNoisyData(model, n, np, p;
+        vdata, xsol, outliers = generate_noisy_data(model, n, np, p;
                                    tMin=tMin, tMax=tMax, xSol=xSol,
                                    std=std, outTimes=outTimes)
     
@@ -124,14 +124,14 @@ end
 
 """
 
-    generateNoisyData(model::Function, n::Int, np::Int, p::Int;
+    generate_noisy_data(model::Function, n::Int, np::Int, p::Int;
                       tMin::Float64=-10.0, tMax::Float64=10.0,
                       xSol::Vector{Float64}=10.0 * randn(Float64, n),
                       std::Float64=200.0, outTimes::Float64=7.0)
 
-    generateNoisyData(model::Function, n, np, p, tMin::Float64, tMax::Float64)
+    generate_noisy_data(model::Function, n, np, p, tMin::Float64, tMax::Float64)
 
-    generateNoisyData(model::Function, n::Int, np::Int, p::Int,
+    generate_noisy_data(model::Function, n::Int, np::Int, p::Int,
                       xSol::Vector{Float64}, tMin::Float64, tMax::Float64)
 
 Random generate a fitting one-dimensional data problem.
@@ -152,7 +152,7 @@ It returns a tuple `(data, xSol, outliers)` where
   - `outliers`: the outliers of this data set
 
 """
-function generateNoisyData(model::Function, n::Int, np::Int, p::Int;
+function generate_noisy_data(model::Function, n::Int, np::Int, p::Int;
                            tMin::Float64=-10.0, tMax::Float64=10.0,
                            xSol::Vector{Float64}=10.0 * randn(Float64, n),
                            std::Float64=200.0, outTimes::Float64=7.0)
@@ -189,9 +189,9 @@ function generateNoisyData(model::Function, n::Int, np::Int, p::Int;
 
 end
 
-generateNoisyData(model::Function, n::Int, np::Int, p::Int, tMin::Float64, tMax::Float64) =
-    generateNoisyData(model, n, np, p; tMin=tMin, tMax=tMax)
+generate_noisy_data(model::Function, n::Int, np::Int, p::Int, tMin::Float64, tMax::Float64) =
+    generate_noisy_data(model, n, np, p; tMin=tMin, tMax=tMax)
 
-generateNoisyData(model::Function, n::Int, np::Int, p::Int,
+generate_noisy_data(model::Function, n::Int, np::Int, p::Int,
                   xSol::Vector{Float64}, tMin::Float64, tMax::Float64) =
-                      generateNoisyData(model, n, np, p; tMin=tMin, tMax=tMax, xSol=xSol)
+                      generate_noisy_data(model, n, np, p; tMin=tMin, tMax=tMax, xSol=xSol)
