### 2019-04-15 1eb645a
Remove duplicate function definition from documentation.
