+++ dev/search_index.js
@@ -137,19 +137,19 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "api/#RAFF.setRAFFOutputLevel",
+    "location": "api/#RAFF.set_raff_output_level",
     "page": "API",
-    "title": "RAFF.setRAFFOutputLevel",
+    "title": "RAFF.set_raff_output_level",
     "category": "function",
-    "text": "setRAFFOutputLevel(level::LogLevel)\n\nSet the output level of raff and praff algorithms to the desired logging level. Options are (from highly verbose to just errors): Logging.Debug, Logging.Info, Logging.Warn and Logging.Error. The package Logging needs to be loaded.\n\nDefaults to Logging.Error.\n\n\n\n\n\n"
+    "text": "set_raff_output_level(level::LogLevel)\n\nSet the output level of raff and praff algorithms to the desired logging level. Options are (from highly verbose to just errors): Logging.Debug, Logging.Info, Logging.Warn and Logging.Error. The package Logging needs to be loaded.\n\nDefaults to Logging.Error.\n\n\n\n\n\n"
 },
 
 {
-    "location": "api/#RAFF.setLMOutputLevel",
+    "location": "api/#RAFF.set_lm_output_level",
     "page": "API",
-    "title": "RAFF.setLMOutputLevel",
+    "title": "RAFF.set_lm_output_level",
     "category": "function",
-    "text": "setLMOutputLevel(level::LogLevel)\n\nSet the output level of lmlovo algorithm to the desired logging level. Options are (from highly verbose to just errors): Logging.Debug, Logging.Info, Logging.Warn and Logging.Error. The package Logging needs to be loaded.\n\nDefaults to Logging.Error.\n\n\n\n\n\n"
+    "text": "set_lm_output_level(level::LogLevel)\n\nSet the output level of lmlovo algorithm to the desired logging level. Options are (from highly verbose to just errors): Logging.Debug, Logging.Info, Logging.Warn and Logging.Error. The package Logging needs to be loaded.\n\nDefaults to Logging.Error.\n\n\n\n\n\n"
 },
 
 {
@@ -157,7 +157,47 @@ var documenterSearchIndex = {"docs": [
     "page": "API",
     "title": "Main functions",
     "category": "section",
-    "text": "lmlovo\nraff\npraff\nsetRAFFOutputLevel\nsetLMOutputLevel"
+    "text": "lmlovo\nraff\npraff\nset_raff_output_level\nset_lm_output_level"
+},
+
+{
+    "location": "api/#RAFF.eliminate_local_min!",
+    "page": "API",
+    "title": "RAFF.eliminate_local_min!",
+    "category": "function",
+    "text": "eliminate_local_min!(sols::Vector{RAFFOutput})\n\nCheck if the function value of the solution found by smaller values of p is not greater when compared with larger ones. This certainly indicates that a local minimizer was found by the smaller p.\n\n\n\n\n\n"
+},
+
+{
+    "location": "api/#RAFF.sort_fun!",
+    "page": "API",
+    "title": "RAFF.sort_fun!",
+    "category": "function",
+    "text": "This function is an auxiliary function. It finds the p smallest values of vector V and brings them to the first p positions. The indexes associated with the p smallest values are stored in ind.\n\n\n\n\n\n"
+},
+
+{
+    "location": "api/#RAFF.update_best",
+    "page": "API",
+    "title": "RAFF.update_best",
+    "category": "function",
[…]
