### 2019-04-15 fsobral/RAFF.jl#11 @fsobral
The `examples` directory came from old versions of `RAFF` and will be moved to `tests`, since some tests use its files.

The `script` directory is the place where we insert scripts for specific tests, conceptual examples and problem generation and visualization. They definitely should stay up-to-date. I'll work on that issue.
