### 2019-04-15 61f9a52
Update new names for random generation functions. #10.
