+++ v0.4.3/search_index.js
@@ -1,0 +1,259 @@
+var documenterSearchIndex = {"docs": [
+
+{
+    "location": "#",
+    "page": "Overview",
+    "title": "Overview",
+    "category": "page",
+    "text": ""
+},
+
+{
+    "location": "#Overview-1",
+    "page": "Overview",
+    "title": "Overview",
+    "category": "section",
+    "text": "This page is devoted to document and make easier the use of RAFF- Robust Algebraic Fitting Function. Our intent is to provide a package to determine fitting functions for a dataset with ability to detect possible outliers of the dataset. All the code was made in Julia language, version 1.0.This package is not an implentation of classical least square solvers. In fact, RAFF is a solver for Low Order Value Problem [1] which is a generalization of least square problems."
+},
+
+{
+    "location": "#Current-Status-1",
+    "page": "Overview",
+    "title": "Current Status",
+    "category": "section",
+    "text": "The current status of this project is beta quality, don\'t use for anything important.  We provide support to serial and parallel running."
+},
+
+{
+    "location": "#Developed-by-1",
+    "page": "Overview",
+    "title": "Developed by",
+    "category": "section",
+    "text": "This project was developed by the optimization group at Department of Mathematics, State University of Maringá, BrazilFrancisco Sobral (Leader)\nEmerson Vitor Castelani\nRonaldo Lopes\nWesley Shirabayashi"
+},
+
+{
+    "location": "#References-1",
+    "page": "Overview",
+    "title": "References",
+    "category": "section",
+    "text": "[1] Andreani, R., Martínez, J. M., Martínez, L., & Yano, F. S. (2009). Low order-value  optimization and applications. Journal of Global Optimization, 43(1), 1-22."
+},
+
+{
+    "location": "tutorial/#",
+    "page": "Tutorial",
+    "title": "Tutorial",
+    "category": "page",
+    "text": ""
+},
+
+{
+    "location": "tutorial/#Tutorial-1",
+    "page": "Tutorial",
+    "title": "Tutorial",
+    "category": "section",
+    "text": ""
+},
+
+{
+    "location": "tutorial/#Installation-1",
+    "page": "Tutorial",
+    "title": "Installation",
+    "category": "section",
+    "text": "This package is supported just for Julia version 1.0. Consequently,  it uses package 3.0. Currently RAFF is not in  Metadata.jl, so the  package can be installed with the Julia package manager. From the Julia REPL, type ] to enter into Pkg REPL mode and run:pkg> add https://github.com/fsobral/RAFF.jl"
+},
+
+{
+    "location": "tutorial/#Basic-usage-1",
+    "page": "Tutorial",
+    "title": "Basic usage",
+    "category": "section",
[…]
