### 2019-04-16 fsobral/RAFF.jl#11 @matbesancon
Ok great. When you move the `examples`, add to the documentation and README that examples can be found in the tests, and distinguish unit tests from these examples if you can
