### 2019-04-16 c28304b
included Julia 1.1 version in travis.yml
