+++ .travis.yml
@@ -7,6 +7,7 @@ os:
 julia:
 #  - 0.7
   - 1.0
+  - 1.1  
 #  - nightly
 
 # # Uncomment the following lines to allow failures on nightly julia
