### 2019-04-16 88dc431
style changes and tutorial installation updated
