### 2019-04-17 86793da
Update README.MD
Update readme for simplified read of installation
