### 2019-04-17 37e468b
Remove unused files. #11
