### 2019-04-17 278296f
Add Multivariate example. #11
