### 2019-04-17 7d51a43
Add information about the 'examples/' directory. #11
