### 2019-04-17 cb6c2c4
Add section 'Examples' to documentation. #11
