### 2019-04-17 e11eee5
Add information regarding 'Examples'. #11
