### 2019-04-17 235071b
Add information regarding the test files and the scripts. #11
