+++ docs/make.jl
@@ -6,7 +6,8 @@ makedocs(
     pages = ["Overview" => "index.md",
 	     "Tutorial"=> "tutorial.md",
 	     "Examples" => "examples.md",
-             "API" => "api.md"],
+             "API" => "api.md",
+             "Advanced" => "advanced.md"],
     #html_prettyurls = false
     #format = Documenter.HTML(prettyurls = false),
     modules = [RAFF]
