### 2019-04-17 8706ac3
Fixed script 'generate_fit_tests'. #11
