+++ dev/search_index.js
@@ -61,7 +61,7 @@ var documenterSearchIndex = {"docs": [
     "page": "Tutorial",
     "title": "Installation",
     "category": "section",
-    "text": "This package is supported just for Julia version 1.0. Consequently,  it uses package 3.0. Currently RAFF is registered in General Julia Registers, so the  package can be installed using the Julia package manager. From the Julia REPL, type ] to enter into Pkg REPL mode and run:pkg> add RAFF"
+    "text": "This package is supported just for Julia version 1.0. Consequently,  it uses package 3.0. Currently RAFF is registered in General Julia Registers, so the  package can be installed using the Julia package manager. From the Julia REPL, type ] to enter into Pkg REPL mode and run:pkg> add RAFFIn what follows, we provide some simple examples on how to solve problems with RAFF. All the examples, and some other ones, are given in the examples/ directory as Julia scripts."
 },
 
 {
@@ -69,7 +69,7 @@ var documenterSearchIndex = {"docs": [
     "page": "Tutorial",
     "title": "Basic usage",
     "category": "section",
-    "text": "Just to illustrate the potential and basic usage of RAFF, let us consider the following data set given by an array A:A=left beginarraycc\n -20   50 \n -15   325\n -10   20 \n -05   125\n  00   10 \n  05   125\n  10   20 \n  15   325\n  20   50 \nendarrayrightLet\'s suppose the first column of A as an experimental measure with  result given by second column. It is easy to see that the fitting  function in this case is accurate and given by phi(t)=t^2 +1Now let\'s perturb one result of second column of A. For example,  consider A_62 = 255. Assuming the model for fitting given byvarphi(xt)=x_1 t^2 +x_2 we have by classical least squares as result x = [0.904329, 1.3039]. On the other hand, when we consider the RAFF algorithm we obtain the correct answer x=[1.0, 1.0]. Moreover, we also have a list of the possible outliers.In order to run RAFF algorithm we need to setup using RAFFand define the data set and model:A=[-2.0  5.0; \n  -1.5  3.25;\n  -1.0  2.0 ;\n  -0.5  1.25;\n   0.0  1.0 ;\n   0.5  2.55;\n   1.0  2.0 ;\n   1.5  3.25;\n   2.0  5.0 ;]\n\nmodel(x, t) = x[1] * t[1]^2 + x[2]After that, we can run method raff:raff(model, A, 2)The number 2 above is the number of variables in model, i. e., the number of parameters to adjust in the model. The output is a RAFFOutput type. For example to access only the parameters of solution, we can useoutput = raff(model, A, 2)\noutput.solutionNote that RAFF algorithm detects and ignores possible outliers. In order to see which points are outliers, we can access the outliers attribute.output.outliersMore details about RAFFOutput type and other options can be obtained in API section.By default RAFF uses automatic differentiation, more specifically ForwardDiff package. But is possible to call RAFF methods with gradient vector of model. For example, considering the above example, we have,nabla varphi(x t) = t^2 1Programming this gradient and run raff we havegmodel(x, t, g)=begin\n   g[1] = t[1]^2\n   g[2] = 1.0\n   return g\nend\n\nraff(model, gmodel, A, 2)Preliminary tests have shown that the use of explicit derivatives is 10 times faster than automatic differentiation."
[…]
