### 2019-04-18 b10578f
change RAFFOutput
