### 2019-04-18 ab6b4a8
fixed some typos of raffoutput
