+++ src/raffoutput.jl
@@ -57,7 +57,7 @@ function show(io::IO, ro::RAFFOutput)
 	print(io,"Status (.status) = $(ro.status) \n")
 	print(io,"Solution (.solution) = $(ro.solution) \n")
 	print(io,"Number of iterations (.iter) = $(ro.iter) \n")
-	print(io,"Number of outliers (.p) = $(ro.p) \n")
-	print(io,"Objetive function value (.f) = $(ro.f) \n")
+	print(io,"Number of trust points (.p) = $(ro.p) \n")
+	print(io,"Objective function value (.f) = $(ro.f) \n")
 	print(io,"Index of outliers (.outliers)= $(ro.outliers)\n")
 end
