### 2019-04-18 1aa26b8
Add 'example' and 'script' explanation to the README. #11
