### 2019-04-18 fsobral/RAFF.jl#11 @fsobral
Dear @matbesancon . I think we managed to organize the examples and scripts:

  - Inside `examples` we created some scripts to run small examples of `RAFF`. In particular, the examples form the tutorial were implemented.
  - All the test files inside `examples` were moved to `test`, since they were only used inside the unit tests.
  - The `script` directory has also been moved to `test` and an explanation of each file has been added to the README and documentation.

Thanks for the suggestion.
