### 2019-04-18 fsobral/RAFF.jl#11 @matbesancon
That's great. You might want to make your modifications in pull requests in different branches, and then merge them to master. That way you can check that Travis CI passes before committing to master. This also groups commits together
