### 2019-04-24 1f8ded5
Fix performance tests and add documentation for the script.
