### 2019-04-29 fsobral/RAFF.jl#13 @matbesancon
[JOSS] Merge JOSS branch
The structure of the `JOSS` branch is very different from the normal repo. The `joss` branch should have the same structure as the master branch (that of a Julia package), with a `paper` folder containing the paper.md and .bib files.

This issue is raised as part of: https://github.com/openjournals/joss-reviews/issues/1385
