### 2019-04-29 fsobral/RAFF.jl#14 @matbesancon
[JOSS] add contribution guidelines
The authors should add some general guidelines for people who want to contribute by either submitting PRs or issues. See https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements
How do people reach out? On the repo or on another platform?

This issue is raised as part of https://github.com/openjournals/joss-reviews/issues/1385
