### 2019-05-02 fsobral/RAFF.jl#15 @abelsiqueira
[JOSS] DOI
The following DOI is missing:

https://link.springer.com/article/10.1007%2Fs11750-010-0169-1

Related to https://github.com/openjournals/joss-reviews/issues/1385
