### 2019-05-02 fsobral/RAFF.jl#16 @abelsiqueira
[JOSS] A statement of need on documentation
I think that the documentation could benefit from more information in the overview. Something like the following part of the second paragraph from the Summary section of the paper:

RAFF provides automatic detection of outliers using a voting system. It is an optimization-based package, based on algorithms for Lower Order-Value Optimization (LOVO) which were introduced in [@Andreani2005] and revisited in [@Andreani2009] to fit the user-provided models $\phi$ to experimental data. Recently, a good review can be found in [@Martinez2012]. In order to find a robust adjustment, a voting system is used, which is also responsible for the detection of possible outliers.

Related to https://github.com/openjournals/joss-reviews/issues/1385
