### 2019-05-02 847415d
Add files from master to branch joss. #13
