### 2019-05-02 fsobral/RAFF.jl#13 @fsobral
Solved!
