+++ paper.bib
@@ -33,6 +33,7 @@ volume={20},
 number={1},
 pages={75--98},
 year={2012},
+doi = {https://doi.org/10.1007/s11750-010-0169-1}
 publisher={Springer}
 }
 
