### 2019-05-02 4fbe3b6
minor corrections in references
