### 2019-05-03 fsobral/RAFF.jl#15 @fsobral
Done.
