### 2019-05-03 fsobral/RAFF.jl#16 @fsobral
Done. It is currently available in the [`dev` documentation](https://fsobral.github.io/RAFF.jl/dev/).
