### 2019-05-03 36ebd78
Add 'Contributions' in README as asked in #14.
