+++ README.MD
@@ -59,6 +59,20 @@ using `RAFF`. Those files can be found in
 script file, please refer to the
 [documentation](https://fsobral.github.io/RAFF.jl/stable/advanced/).
 
+## Contributing
+
+Any suggestions or contributions to `RAFF` are very welcome. `RAFF` is
+a open-source software available through the MIT license, so users are
+free to create their own versions of the software. Users are specially
+encouraged to participate on the project in the following ways:
+
+  - Modifications to the code can be submitter as pull requests to the
+	branch `master`.
+  - Bugs, suggestions and questions can be raised as issues in the
+    project's Github page.
+  - General doubts can also be sent directly to `fncsobral at uem dot
+    br`.
+
 ## Credits
 
 * Francisco Sobral (Leader)
