### 2019-05-03 c74ed56
New version containing the new round of suggestions by the referees.
