### 2019-05-03 fsobral/RAFF.jl#14 @fsobral
Thank you for the suggestion. I have added a 'Contribution' section in the `README`.
