### 2019-05-06 fsobral/RAFF.jl#17 @matbesancon
[JOSS] Gradient function signature
From the documentation tutorial:
```
julia> gmodel(x, t, g)=begin
          g[1] = t[1]^2
          g[2] = 1.0
          return g
       end
```

Following the Julia style guide, the function could have a bang `gmodel!` to indicate it is
modifying its arguments. Moreover, the mutated argument is preferred placed first, as in `gmodel!(g, x, t)`

Issue created for https://github.com/openjournals/joss-reviews/issues/1385
