### 2019-05-06 fsobral/RAFF.jl#18 @matbesancon
[JOSS] Paper review 1
You will find below my initial review of the paper itself:

Title of the package:
Why "Algebraic"? It uses numerical methods and not symbolic differentiation.

First line: use `\mathbb{R}` instead of `\mathcal{R}`

Prefer `Package.jl` to Package to designate it

Second paragraph of "Summary"
"adjustment of mathematical functions to data" is also refered to as regression, which is not mentioned in the paper, while it is fundamentally the purpose of the package.

"Detection of outliers is always regarded ...". It's hard to see what the authors mean here. Without proper references backing up the statement, this seems rather vague.

"Recently, a good review", the use of "good" is subjective (what is a good review, in what context). A review of what?

In statistics, it is common to use f(x; θ) for a function of x with parameters theta, which one tries to fit in a regression task.

In paragraph "Functionality", the equation is not homogeneous with the rest:
`ϕ(x,t)` instead of `ϕ(x;t)`, prefer ";" to indicate the separation of data and parameters.

There could be a section on the method itself, or at least more development, including on the LOVO algorithm. For instance, the method is not black-box optimization but assumes the existence of gradient, thus making it first-order. Furthermore, if gradients are not provided by users, the use of `ForwardDiff` makes it mandatory to have the function defined in a generic way (accepting any `Real` number).
Thus, how is the proposed method different from using other first-order methods (such as the ones in Optim.jl or JuliaSmoothOptimizers)? This should be detailed.

"it is obvious that its arguments [...]", avoid the use of "obvious" which is subjective, and develop what is meant here.

"Additional features" paragraph, sentence: "the distributed version is a centralized ..." is a contradiction and should be re-phrased.

This belongs to the JOSS review https://github.com/openjournals/joss-reviews/issues/1385
