### 2019-05-06 #1385 @matbesancon
@abelsiqueira I did a first review of the paper at https://github.com/fsobral/RAFF.jl/issues/18 if you want to add some comments directly to it
