### 2019-05-06 fsobral/RAFF.jl#18 @evcastelani
1. [x] Title of the package: Why "Algebraic"? It uses numerical methods and not symbolic differentiation.

    The term "algebraic" refers to the type of problem solved and not to the programming techniques of the algorithm. Although it is not a common term in adjustment functions, it serves to differ from the geometric fit (geometrical fit functions). In the case of geometric fit, we are interested in minimizing the sum of deviations of the type  `|| phi (t, x) - (t_i, y_i) ||` (ie, orthogonal projection on the curve to be adjusted). In the algebraic case, we are interested in minimizing the sum of deviations of type `|| phi (t_i, x) - y_i ||`. We chose this term to make the package's name more readable.

1. [ ] First line: use \mathbb{R} instead of \mathcal{R}

    TODO: @evcastelani

1. [ ] Prefer Package.jl to Package to designate it

    TODO: @evcastelani

1. [ ] Second paragraph of "Summary". "adjustment of mathematical functions to data" is also referred to as regression, which is not mentioned in the paper, while it is fundamentally the purpose of the package.

   * Just change file *
   TODO: @fsobral

1. [ ] "Detection of outliers is always regarded ...". It's hard to see what the authors mean here. Without proper references backing up the statement, this seems rather vague.

  - Mention the statistical techniques to detect and remove outliers
  - Mention paper
  - TODO: @fsobral

1. [ ] "Recently, a good review", the use of "good" is subjective (what is a good review, in what context). A review of what?

TODO: @evcastelani

1. [ ] In statistics, it is common to use f(x; θ) for a function of x with parameters theta, which one tries to fit in a regression task.

    TODO: @fsobral

1. [ ] In paragraph "Functionality", the equation is not homogeneous with the rest:
ϕ(x,t) instead of ϕ(x;t), prefer ";" to indicate the separation of data and parameters.

    TODO: @fsobral

1. [ ] There could be a section on the method itself, or at least more development, including on the LOVO algorithm. For instance, the method is not black-box optimization but assumes the existence of gradient, thus making it first-order. Furthermore, if gradients are not provided by users, the use of ForwardDiff makes it mandatory to have the function defined in a generic way (accepting any Real number).
Thus, how is the proposed method different from using other first-order methods (such as the ones in Optim.jl or JuliaSmoothOptimizers)? This should be detailed.

1. [ ] "it is obvious that its arguments [...]", avoid the use of "obvious" which is subjective, and develop what is meant here.

    TODO: @evcastelani

1. [ ] "Additional features" paragraph, sentence: "the distributed version is a centralized ..." is a contradiction and should be re-phrased.

    TODO: @fsobral
