### 2019-05-07 fsobral/RAFF.jl#18 @matbesancon
@evcastelani thanks for the answer list. For the algebraic part, putting this explanation and/or references in the paper would be useful
