### 2019-05-07 fsobral/RAFF.jl#18 @matbesancon
ping me when the editions are done
