### 2019-05-07 b618f51
Change signature, comments and variables for `lmlovo`.
