### 2019-05-08 300bba3
Change vector of parameters to \theta in 'lmlovo'
