+++ src/RAFF.jl
@@ -30,10 +30,10 @@ include("dutils.jl")
 include("generator.jl")
 
 """
-    lmlovo(model::Function [, t::Vector{Float64} = zeros(n)], data::Array{Float64, 2},
+    lmlovo(model::Function [, θ::Vector{Float64} = zeros(n)], data::Array{Float64, 2},
            n::Int, p::Int [; kwargs...])
 
-    lmlovo(model::Function, gmodel!::Function [, t::Vector{Float64} = zeros(n)],
+    lmlovo(model::Function, gmodel!::Function [, θ::Vector{Float64} = zeros(n)],
            data::Array{Float64,2}, n::Int, p::Int [; MAXITER::Int=200,
            ε::Float64=10.0^-4])
 
@@ -51,23 +51,23 @@ Matriz `data` is the data to be fit. This matrix should be in the form
 where `N` is the dimension of the argument of the model
 (i.e. dimension of `t`).
 
-If 't' is provided, the it is used as the starting point.
+If 'θ' is provided, then it is used as the starting point.
 
 The signature of function `model` should be given by
 
-    model(x::Union{Vector{Float64}, SubArray}, t::Vector{Float64})
+    model(x::Union{Vector{Float64}, SubArray}, θ::Vector{Float64})
 
-where `x` are the variables and `t` is a `n`-dimensional vector of
+where `x` are the variables and `θ` is a `n`-dimensional vector of
 parameters. If the gradient of the model `gmodel!`
 
     gmodel!(g::Vector{Float64}, x::Union{Vector{Float64},
-            t::Vector{Float64}, SubArray} )
+            θ::Vector{Float64}, SubArray} )
 
 is not provided, then the function ForwardDiff.gradient! is called to
 compute it.  **Note** that this choice has an impact in the
 computational performance of the algorithm. In addition, if
 `ForwardDiff.jl` is being used, then one **MUST** remove the signature
-of vector `t` from the model.
+of vector `θ` from the model.
 
 The optional arguments are
 
@@ -77,7 +77,7 @@ The optional arguments are
 Returns a [`RAFFOutput`](@ref) object.
 
 """
-function lmlovo(model::Function, gmodel!::Function, t::Vector{Float64},
+function lmlovo(model::Function, gmodel!::Function, θ::Vector{Float64},
                 data::Array{Float64,2}, n::Int, p::Int;
                 MAXITER::Int=200, ε::Float64=10.0^-4)
 
@@ -92,7 +92,7 @@ function lmlovo(model::Function, gmodel!::Function, t::Vector{Float64},
 
     end
 
-    (p == 0) && return RAFFOutput(1, t, 0, p, 0.0, [1:npun;])
+    (p == 0) && return RAFFOutput(1, θ, 0, p, 0.0, [1:npun;])
     
     # Main function - the LOVO function
     LovoFun = let
@@ -106,10 +106,10 @@ function lmlovo(model::Function, gmodel!::Function, t::Vector{Float64},
         p_::Int = p
         
         # Return a ordered set index and lovo value
-        (t) -> begin
+        (θ) -> begin
             
             @views for i = 1:npun_
-                F[i] = (model(data[i,1:(end - 1)]) - data[i, end], t)^2
+                F[i] = (model(data[i,1:(end - 1)]) - data[i, end], θ)^2
             end
             
             indF, orderedF = sort_fun!(F, ind, p_)
@@ -125,18 +125,18 @@ function lmlovo(model::Function, gmodel!::Function, t::Vector{Float64},
     jac_res::Array{Float64, 2} = Array{Float64}(undef, p, n)
     
     # This function returns the residue and Jacobian of residue
[…]
