### 2019-05-08 bdf59ed
Change vector of parameters to \theta in 'raff'. Fix docstrings.
