### 2019-05-08 22d884e
Change signatures in 'praff' and 'dutils', the distributed case.
