### 2019-05-08 9f3f0a4
Update tests in order to change signature of the model.
