### 2019-05-08 79c2203
Update correct signature of 'gmodel'.
