### 2019-05-08 0409a30
Update examples with new signatures.
