### 2019-05-08 e5efb92
Update version. Major changes, since the signature of 'model' and 'gmodel\!' has ben changed.
