### 2019-05-08 fsobral/RAFF.jl#17 @fsobral
@matbesancon We have fixed this issue in the documentation and in the whole code as well. The version has been updated to v0.6.0, since all the signatures of `gmodel` had to be changed.
