### 2019-05-08 fsobral/RAFF.jl#13 @fsobral
Dear @matbesancon . Although we have made this change, now the RAFF version in the branch `joss` does not follow the `master` branch. To do so, we will always have to merge it manually into `joss`, so it becomes up-to-date. Do you have a better alternative to this? We think that there is no sense to have a `joss` directory inside the package's files, but that could be a solution...
