### 2019-05-08 fsobral/RAFF.jl#18 @abelsiqueira
## Review

The authors implement a solver for data fitting that automatically detects outliers, based on algorithms using Lower Order-Value Optimization. The package meets the criteria for publication on JOSS, but there are a few items that should be addressed, and a few suggestions.

- What kind of optimal fitting is achieved? Seems to be least squares (for the data without the outliers), but it's worth mentioning in the summary;
- In what sense are the outliers defined? Some quantile?
- Since the problem in the example is nonlinear, it should be indicated which solver was used to solve the problem, and which starting point;
- I think that a comparison with a different fitting not so sensitive to outliers would improve the quality of the paper. Perhaps fitting with norm-1 or Huber loss.
- The releases are not being registered. The Julia register has 0.4.3, but there is a release 0.6.0 already and the docs are being updated according to the releases;
- In the first paragraph of **Additional features**: "...provided, what results in...", change "what" to "which".
- On Figure 1, move the legend to the right, use a stronger color for the data points that are not outliers, and make it black-and-white/color-blind friendly by using a different marker style for the outliers and changing line-styles (I'd remove the "True solution").

Suggestions: These can be disregarded it the authors prefer:

- In the context of regression, I find that it's clearer to use `x` and `y` for the data and `\beta` or `\theta` for the parameters being optimized. Our traditional least squares problem `min |Ax - b|^2` is usually written as `min |Xβ - y|^2`.
- I find that entering the data separately for `x` and `y` is better, e.g.
```
x = range(-1, 1, length=100)
y = 2 .+ 3x + randn(100) * 0.1
model(β, x) = β[1] + β[2] * x[1]
output = raff(model, x, y, 2) # Here
```
