+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'RAFF: Robust Algebraic Fitting Function in Julia'
+title: 'RAFF.jl: Robust Algebraic Fitting Function in Julia'
 tags:
   - Julia
   - Statistics
@@ -30,48 +30,111 @@ bibliography: paper.bib
 Let $f : \mathbb{R}^n \to \mathbb{R}$ be a function whose mathematical description
 is not available. This function can be, for example, a black-box, a
 proprietary computer program or an experiment. Suppose that a data set
-$S = \{(t_1, f(t_1)), \dots, (t_m, f(t_m))\}$ of $m$ observations of
+$S = \{(x_1, f(x_1)), \dots, (x_m, f(x_m))\}$ of $m$ observations of
 $f$ is available and we want to approximate $f$ by a known model
-$\phi$. Model $\phi$ can be defined as $\phi(x;t)$, where $t$ are the
-$n$ independent variables of $f$ and $x$ represents some parameters of
-$\phi$. ``RAFF`` (Robust Algebraic Fitting Function) is a Julia
-package developed to find the optimal parameters $x$ for $\phi$ in
+$\phi$. Model $\phi$ can be defined as $\phi(x;\theta)$, where $x$ are the
+$n$ independent variables of $f$ and $\theta$ represents some parameters of
+$\phi$. ``RAFF.jl`` (Robust Algebraic Fitting Function) is a Julia
+package developed to find the optimal parameters $\theta$ for $\phi$ in
 order to adjust it to the observed values $S$ of the unknown function
-$f$.
+$f$. Following [@Liu2008], in general, the adjust can be related to 
+
+1. Classical least square (algebraic fit): which consider the sum of deviations of type
+$\vert \phi(x_i; \theta) - f(x_i) \vert^2$;
+2. Ortogonal least square (geometric fit): which consider the sum of deviations of type
+$\Vert \phi(x;\theta)-(x_i,f(x_i))\Vert^2$ (orthogonal projection on the
+curve to be adjusted).
+
+``RAFF.jl`` was developed to solve a generalization of the first case.
 
 The adjustment of mathematical functions to data is a problem that
 appears in many areas of science. When data comes from real
 experiments, non-expected errors may cause the appearance of
 outliers. Detection of outliers is always regarded as the statistical
-part of data adjustment. ``RAFF`` provides automatic detection of
+part of data adjustment. ``RAFF.jl`` provides automatic detection of
 outliers using a voting system. It is an optimization-based package,
 based on algorithms for Lower Order-Value Optimization (LOVO)
 which were introduced in [@Andreani2005] and revisited in [@Andreani2009] 
 to fit the user-provided models $\phi$ to experimental
-data. Recently, a good review can be found in [@Martinez2012]. In order to 
-find a robust adjustment, a voting system is used,
-which is also responsible for the detection of possible outliers.
+data. Recently, a complete review about LOVO problems considering 
+theoretical aspects of algorithms to solve it and potential applications 
+can be found in [@Martinez2012]. In order to find a robust adjustment, a 
+voting system is used, which is also responsible for the detection of 
+possible outliers.
+
+# Background Material
+
+To elucidate the essence of how ``RAFF.jl`` works, let us detail some aspects related
+to the LOVO problem and its resolution. Let us consider $m$ functions
+$F_i:\mathbb{R}^n \rightarrow \mathbb{R}$, $i=1,...,m$. Given $\theta \in
[…]
