+++ joss/paper.bib
@@ -33,7 +33,7 @@ volume={20},
 number={1},
 pages={75--98},
 year={2012},
-doi = {https://doi.org/10.1007/s11750-010-0169-1}
+doi = {10.1007/s1175001001691},
 publisher={Springer}
 }
 
@@ -43,4 +43,28 @@ publisher={Springer}
   journal = {arXiv:1607.07892 [cs.MS]},
      year = {2016},
      url = {https://arxiv.org/abs/1607.07892}
+}
+
+@InProceedings{Liu2008,
+author={Liu, Yang and Wang, Wenping},
+editor={Chen, Falai and J{\"u}ttler, Bert},
+title="A Revisit to Least Squares Orthogonal Distance Fitting of Parametric Curves and Surfaces",
+booktitle={Advances in Geometric Modeling and Processing},
+year={2008},
+publisher={Springer Berlin Heidelberg},
+address={Berlin, Heidelberg},
+pages={384--397},
+url={https://doi.org/10.1007/978-3-540-79246-8_29},
+isbn={978-3-540-79246-8}
+}
+
+@article{more1978levenberg,
+title={The Levenberg-Marquardt algorithm: Implementation and theory},
+author={Mor{\'e}, Jorge},
+journal={Numerical Analysis},
+pages={105--116},
+year={1978},
+publisher={Springer},
+doi={10.1007/BFb0067700},
+url={https://doi.org/10.1007/BFb0067700}
 }
