### 2019-05-08 fsobral/RAFF.jl#18 @fsobral
- [ ] What kind of optimal fitting is achieved? Seems to be least squares (for the data without the outliers), but it's worth mentioning in the summary;

    Maybe the question has already been answered in the new version.
    TODO: @evcastelani

- [ ] In what sense are the outliers defined? Some quantile?

    Maybe the question has already been answered in the new version.

- [ ] Since the problem in the example is nonlinear, it should be indicated which solver was used to solve the problem, and which starting point;

    - Explicitly state the problem solved by `RAFF`.
    - Explain better here.

- [ ] I think that a comparison with a different fitting not so sensitive to outliers would improve the quality of the paper. Perhaps fitting with norm-1 or Huber loss.

    We do not think that the 1-norm would be more robust than the least squares (suppose that the errors are smaller than 1, for example). In addition, the problem would become non-differentiable and harder to solve. Moreover, one can always create a sufficiently far outlier so that the "more robust" approaches also will construct poor models. Our example was intended to show how could one benefit for *detecting* and *eliminating* outliers, which is different from simply tolerating them.

- [ ] The releases are not being registered. The Julia register has 0.4.3, but there is a release 0.6.0 already and the docs are being updated according to the releases;

  - Is there a way of doing automatically?

- [ ] In the first paragraph of **Additional features**: "...provided, what results in...", change "what" to "which".

- [ ] On Figure 1, move the legend to the right, use a stronger color for the data points that are not outliers, and make it black-and-white/color-blind friendly by using a different marker style for the outliers and changing line-styles (I'd remove the "True solution").

    TODO: @fsobral

Suggestions: These can be disregarded it the authors prefer:

- [x] In the context of regression, I find that it's clearer to use `x` and `y` for the data and `\beta` or `\theta` for the parameters being optimized. Our traditional least squares problem `min |Ax - b|^2` is usually written as `min |Xβ - y|^2`.

- [ ] I find that entering the data separately for `x` and `y` is better, e.g.
    ```
    x = range(-1, 1, length=100)
    y = 2 .+ 3x + randn(100) * 0.1
    model(β, x) = β[1] + β[2] * x[1]
    output = raff(model, x, y, 2) # Here
    ```
