### 2019-05-09 fsobral/RAFF.jl#13 @matbesancon
@fsobral you can manually merge regularly, but the key point is that the structure is now identical (even if joss is a bit late on master).

Also, once the paper is finished, it will not be problematic to merge joss into master
