+++ joss/paper.bib
@@ -44,3 +44,14 @@ publisher={Springer}
      year = {2016},
      url = {https://arxiv.org/abs/1607.07892}
 }
+
+@article{Motulsky2006a,
+author = {Motulsky, Harvey J. and Brown, E. Ronald},
+doi = {10.1186/1471-2105-7-123},
+journal = {BMC Bioinformatics},
+mendeley-groups = {Curve Fitting},
+number = {123},
+title = {{Detecting outliers when fitting data with nonlinear regression – a new method based on robust nonlinear regression and the false discovery rate}},
+volume = {7},
+year = {2006}
+}
