### 2019-05-09 6962641
Add [compat] for julia 1.0, as asked in pull request https://github.com/JuliaRegistries/General/pull/625 for JuliaRegistrator.jl.
