### 2019-05-09 fsobral/RAFF.jl#13 @fsobral
Ok. Thanks for the suggestion.
