+++ joss/paper.md
@@ -31,18 +31,18 @@ Let $f : \mathbb{R}^n \to \mathbb{R}$ be a function whose mathematical descripti
 is not available. This function can be, for example, a black-box, a
 proprietary computer program or an experiment. Suppose that a data set
 $S = \{(x_1, f(x_1)), \dots, (x_m, f(x_m))\}$ of $m$ observations of
-$f$ is available and we want to approximate $f$ by a known model
-$\phi$. Model $\phi$ can be defined as $\phi(x;\theta)$, where $x$ are the
+$f$ are available and we want to approximate $f$ by a known model
+$\phi$. Model $\phi$ can be defined as $\phi(x, \theta)$, where $x$ are the
 $n$ independent variables of $f$ and $\theta$ represents some parameters of
 $\phi$. ``RAFF.jl`` (Robust Algebraic Fitting Function) is a Julia
 package developed to find the optimal parameters $\theta$ for $\phi$ in
 order to adjust it to the observed values $S$ of the unknown function
-$f$. Following [@Liu2008], in general, the adjust can be related to 
+$f$. Following [@Liu2008], in general, the adjustment can be related to 
 
-1. Classical least square (algebraic fit): which consider the sum of deviations of type
-$\vert \phi(x_i; \theta) - f(x_i) \vert^2$;
-2. Ortogonal least square (geometric fit): which consider the sum of deviations of type
-$\Vert \phi(x;\theta)-(x_i,f(x_i))\Vert^2$ (orthogonal projection on the
+1. Classical least squares (algebraic fit): which considers the sum of deviations of type
+$\vert \phi(x_i, \theta) - f(x_i) \vert^2$, also known as regression;
+2. Orthogonal least squares (geometric fit): which considers the sum of deviations of type
+$\Vert \phi(x, \theta)-(x_i,f(x_i))\Vert^2$ (orthogonal projection on the
 curve to be adjusted).
 
 ``RAFF.jl`` was developed to solve a generalization of the first case.
@@ -61,8 +61,7 @@ nonlinear regression based on least squares. ``RAFF.jl`` provides
 automatic detection of outliers using a voting system. It is an
 optimization-based package, based on algorithms for Lower Order-Value
 Optimization (LOVO) which were introduced in [@Andreani2005] and
-revisited in [@Andreani2009] to fit the user-provided models $\phi$ to
-experimental data. Recently, a good review can be found in
+revisited in [@Andreani2009]. Recently, a good review can be found in
 [@Martinez2012]. In order to find a robust adjustment, a voting system
 is used, which is also responsible for the detection of possible
 outliers.
@@ -126,61 +125,40 @@ among them is declared as the solution by ``RAFF.jl``.
 
 # Functionality
 
-``RAFF.jl`` main methods expect as input a data set of the observed data and a
-model function, whose parameters one intends to adjust. The model function is a
-regular Julia function with 2 arguments: $\theta$ represents the parameters of the
-model and $x$ represents the arguments of function $f$. The following function
-is an example of a model representing a circle $$ \phi(x;\theta) = (x_1 - \theta_1)^2 +
-(x_2 - \theta_2)^2 - \theta_3^2.  $$
+``RAFF.jl`` main methods expect as input a data set of the observed
+data and a model function, whose parameters one intends to adjust. The
+model function is a regular Julia function with 2 arguments: $\theta$
+represents the parameters of the model and $x$ represents the
+arguments of function $f$. The following function is an example of a
+model representing the logistic function $$ \phi(x, \theta) =
+\theta_1 + \frac{\theta_2}{1.0 + \exp(- \theta_3 x + \theta_4)}.
+$$ The observed data can be represented by the following table:
 
-Note that $\theta_1$ and $\theta_2$ are parameters related to the center of the circle,
[…]
