+++ joss/paper.md
@@ -27,22 +27,24 @@ bibliography: paper.bib
 
 # Summary
 
-Let $f : \mathbb{R}^n \to \mathbb{R}$ be a function whose mathematical description
-is not available. This function can be, for example, a black-box, a
-proprietary computer program or an experiment. Suppose that a data set
-$S = \{(x_1, f(x_1)), \dots, (x_m, f(x_m))\}$ of $m$ observations of
-$f$ are available and we want to approximate $f$ by a known model
-$\phi$. Model $\phi$ can be defined as $\phi(x, \theta)$, where $x$ are the
-$n$ independent variables of $f$ and $\theta$ represents some parameters of
-$\phi$. ``RAFF.jl`` (Robust Algebraic Fitting Function) is a Julia
-package developed to find the optimal parameters $\theta$ for $\phi$ in
-order to adjust it to the observed values $S$ of the unknown function
-$f$. Following [@Liu2008], in general, the adjustment can be related to 
+Let $f : \mathbb{R}^n \to \mathbb{R}$ be a function whose mathematical
+description is not available. This function can be, for example, a
+black-box, a proprietary computer program or an experiment. Suppose
+that a data set $S = \{(x_1, y_1), \dots, (x_m, y_m)\}$ is available,
+where $y_i$ is an approximation of $f(x_i)$ (from an experimental
+procedure, numerical approximation, etc.) and we want to approximate
+$f$ by a known model $\phi$. Model $\phi$ can be defined as $\phi(x,
+\theta)$, where $x$ are the $n$ independent variables of $f$ and
+$\theta$ represents some parameters of $\phi$. ``RAFF.jl`` (Robust
+Algebraic Fitting Function) is a Julia package developed to find the
+optimal parameters $\theta$ for $\phi$ in order to adjust it to the
+observed values $S$ of the unknown function $f$. Following [@Liu2008],
+in general, the adjustment can be related to
 
 1. Classical least squares (algebraic fit): which considers the sum of deviations of type
-$\vert \phi(x_i, \theta) - f(x_i) \vert^2$, also known as regression;
+$\vert \phi(x_i, \theta) - y_i \vert^2$, also known as regression;
 2. Orthogonal least squares (geometric fit): which considers the sum of deviations of type
-$\Vert \phi(x, \theta)-(x_i,f(x_i))\Vert^2$ (orthogonal projection on the
+$\min_x \Vert (x, \phi(x, \theta))-(x_i, y_i)\Vert^2$ (orthogonal projection on the
 curve to be adjusted).
 
 ``RAFF.jl`` was developed to solve a generalization of the first case.
@@ -61,12 +63,11 @@ nonlinear regression based on least squares. ``RAFF.jl`` provides
 automatic detection of outliers using a voting system. It is an
 optimization-based package, based on algorithms for Lower Order-Value
 Optimization (LOVO) which were introduced in [@Andreani2005] and
-revisited in [@Andreani2009]. Recently, a good review can be found in
-[@Martinez2012]. In order to find a robust adjustment, a voting system
-is used, which is also responsible for the detection of possible
-outliers.
+revisited in [@Andreani2009]. Recently, a complete review about LOVO problems considering 
+theoretical aspects of algorithms to solve it and potential applications 
+can be found in [@Martinez2012].
 
-# Background Material
+# Background
 
 To elucidate the essence of how ``RAFF.jl`` works, let us detail some aspects related
 to the LOVO problem and its resolution. Let us consider $m$ functions
@@ -80,48 +81,58 @@ $$
[…]
