+++ joss/paper.bib
@@ -48,7 +48,7 @@ publisher={Springer}
 
 @article{Motulsky2006a,
 author = {Motulsky, Harvey J. and Brown, E. Ronald},
-url = {10.1186/1471-2105-7-123},
+doi = {10.1186/1471-2105-7-123},
 journal = {BMC Bioinformatics},
 mendeley-groups = {Curve Fitting},
 number = {123},
@@ -66,7 +66,7 @@ year={2008},
 publisher={Springer Berlin Heidelberg},
 address={Berlin, Heidelberg},
 pages={384--397},
-url={https://doi.org/10.1007/978-3-540-79246-8_29},
+doi={10.1007/978-3-540-79246-8_29},
 isbn={978-3-540-79246-8}
 }
 
