### 2019-05-15 d2e653e
added bibliography for OLS
