+++ joss/paper.bib
@@ -69,7 +69,15 @@ pages={384--397},
 doi={10.1007/978-3-540-79246-8_29},
 isbn={978-3-540-79246-8}
 }
-
+@article{keles2018,
+title={Comparison of Classical Least Squares and Orthogonal Regression in Measurement Error Models.},
+author={Keles, Taliha},
+journal={International Online Journal of Educational Sciences},
+volume={10},
+number={3},
+year={2018},
+url={http://www.iojes.net/Makaleler/386bbbe6-b98d-4766-a717-2c945a97f267.pdf}
+}
 @article{more1978levenberg,
 title={The Levenberg-Marquardt algorithm: Implementation and theory},
 author={Mor{\'e}, Jorge},
