### 2019-05-16 fsobral/RAFF.jl#18 @fsobral
Dear @abelsiqueira and @matbesancon , Thank you for you valuable comments and suggestions. We have added responses to most of your comments and questions. If there is no comment, it is because it has been fully accepted.
