### 2019-05-16 fsobral/RAFF.jl#18 @matbesancon
Thank you I should have time to review the changes between now and monday
