### 2019-05-17 fsobral/RAFF.jl#18 @matbesancon
## Second review

### Notation
Page 2: "we can sort the set in ascending order"
The notation below is `F_{i(theta)}(theta)` while in the rest of the paper, it is `F_{i}(theta)`, I didn't see why.
For the notation again, it is counter-intuitive that the functions get re-ordered for every theta value. Because of this, it is hard to see how S_p(theta) is not differentiable (because p is still fixed, but the re-ordering of the indices switches which function is where).

### Method

"RAFF.jl" solves this limitation by implementing a voting system. It is not clear from the end of last paragraph which limitation the authors are referring to.

How is the voting system brute-force?

> "The solution which most occurs"

Solutions are not countable (two solutions are equal iff the functions yield the same image at any in any point in the domain, does that even happen with varying points?)

### Phrasing

`i.e., it solves the problem ...` "i.e." is strangely spaced.

> One of the most usual ways to solve the problem of least-squares

It would be good to emphasize "problem of non-linear least-squares", otherwise pseudo-matrix inversion would be preferred.

"master-slave implementation" --> prefer primary-worker / primary-secondary or other terms. master-slave now tends to be avoided in technical contexts.
