### 2019-05-17 fsobral/RAFF.jl#18 @fsobral
> ## Second review
> ### Notation
>
> Page 2: "we can sort the set in ascending order"
> The notation below is `F_{i(theta)}(theta)` while in the rest of the paper, it is `F_{i}(theta)`, I didn't see why.
> For the notation again, it is counter-intuitive that the functions get re-ordered for every theta value. Because of this, it is hard to see how S_p(theta) is not differentiable (because p is still fixed, but the re-ordering of the indices switches which function is where).

The LOVO function is not very easy to understand. One has to spend some time getting used to this terminology and notation. The easiest way of seeing the non differentiability of S_p(theta) is by observing that f_min(theta) = S_p(theta) for all theta, as discussed in the paragraph starting at line 15 on pg. 2. Also, f_min highlights the combinatorial spirit of the LOVO problem.

Regarding the notation, it is correct. When we write F_{i_1(theta)}(theta) we mean that it is the smallest value in the set {F_{i}(theta), i = 1, ..., m} for a given theta.  F_{i_2(theta)}(theta) is the second smallest value in the set, and so on. The procedure for obtaining this is by evaluating all F_i at theta and ordering the set {F_{i}(theta), i = 1, ..., m}. That's how the indices i_{k}(theta) are obtained. Since the values of F_i change for different values of theta, the order of the F_i's change too.

Therefore, we cannot change the notation, but we will add a small example to explain it better. We will ping you when it is finished.

> ### Method
>
> "RAFF.jl" solves this limitation by implementing a voting system. It is not clear from the end of last paragraph which limitation the authors are referring to.
>
> How is the voting system brute-force?
>
> > "The solution which most occurs"
>
> Solutions are not countable (two solutions are equal iff the functions yield the same image at any in any point in the domain, does that even happen with varying points?)

We have rephrase this paragraph to:

In this voting system, several LOVO subproblems are solved
with different values for $p$, the number of possible trusted
points. Each solution of a LOVO subproblem is associated to a vector
parameter $\theta$. The vector parameters are compared against each
other using the Euclidian distance, where small distances (using a
threshold) are considered the same solution. The parameter $\theta^*$
which most occurs among them is declared as the solution.

> ### Phrasing
>
> `i.e., it solves the problem ...` "i.e." is strangely spaced.
>
> > One of the most usual ways to solve the problem of least-squares
>
> It would be good to emphasize "problem of non-linear least-squares", otherwise pseudo-matrix inversion would be preferred.
>
> "master-slave implementation" --> prefer primary-worker / primary-secondary or other terms. master-slave now tends to be avoided in technical contexts.

Done. Thanks for the observation
