### 2019-05-17 1f7e2f8
Improve paper. #18.
