### 2019-05-17 bc594da
new example to describe LOVO function
