### 2019-05-19 fsobral/RAFF.jl#18 @abelsiqueira
- At the end Functionality, mention that two points were manually set to be outliers.

I'm otherwise satisfied with the updates and the new version.
