### 2019-05-20 fsobral/RAFF.jl#18 @fsobral
>     * At the end Functionality, mention that two points were manually set to be outliers.
>
>
> I'm otherwise satisfied with the updates and the new version.

You are correct. I have added the comment in the same phrase describing of the red triangles.
