### 2019-05-20 9c956fe
Added example for the LOVO function.
