### 2019-05-20 fsobral/RAFF.jl#18 @fsobral
@matbesancon We have added the simple example regarding the LOVO function. I hope it becomes clearer what we mean as its non differentiability.
