### 2019-05-20 b3cdff3
typos corrections
