### 2019-05-21 d85f8d9
minor correction in the text
