### 2019-05-23 fsobral/RAFF.jl#18 @matbesancon
Thanks, it looks good. Minor last thing:

> [...] therefore, can be run locally or in a cluster of computers

Shouldn't it be "on a cluster of computers"?

> used by all experimental researchers who know a little about ...

This sentence looks a bit informal and can be interpreted in different ways: is it for experimental researchers **only if** they know a little, by experimental researchers **even if** they know little about, etc.
