### 2019-05-23 fsobral/RAFF.jl#18 @fsobral
>
> Shouldn't it be "on a cluster of computers"?

Yes, you are right. Done.

>
> > used by all experimental researchers who know a little about ...
>
> This sentence looks a bit informal and can be interpreted in different ways: is it for experimental researchers **only if** they know a little, by experimental researchers **even if** they know little about, etc.

The idea is that they do not need to know much about mathematical modeling, but definitely, they must know something, otherwise writing Julia functions defining models won't make any sense. I have rewritten this sentence to

This package is intended to be used by any experimental researcher
with a little knowledge about mathematical modeling and fitting
functions.
