### 2019-05-23 e72bb1f
Minor corrections to sentences.
