### 2019-05-24 fsobral/RAFF.jl#18 @matbesancon
I'm ok with the changes thanks. Last tiny things in the reference, the letters are not properly capitalizd, you should use `title = {{Title}}` instead of `title = {Title}` everywhere
