+++ joss/paper.bib
@@ -12,7 +12,8 @@ year = {2009}
 }
 
 @article{Andreani2005,
-title={Nonlinear-programming reformulation of the order-value optimization problem},
+title={{Nonlinear-programming reformulation of the order-value optimization
+		problem}},
 author={Andreani, R. and Dunder, C. and Mart{\'\i}nez, J.  M.},
 journal={Mathematical Methods of Operations Research},
 volume={61},
@@ -26,7 +27,7 @@ publisher={Springer}
 }
 
 @article{Martinez2012,
-title={Generalized order-value optimization},
+title={{Generalized order-value optimization}},
 author={Mart{\'\i}nez, J. M.},
 journal={Top},
 volume={20},
@@ -38,7 +39,7 @@ publisher={Springer}
 }
 
 @article{Revels2016,
-    title = {Forward-Mode Automatic Differentiation in Julia},
+    title = {{Forward-Mode Automatic Differentiation in Julia}},
    author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
   journal = {arXiv:1607.07892 [cs.MS]},
      year = {2016},
@@ -47,7 +48,7 @@ publisher={Springer}
 
 
 @article{Motulsky2006a,
-author = {Motulsky, Harvey J. and Brown, E. Ronald},
+author = {Motulsky, H. J. and Brown, E. R. },
 doi = {10.1186/1471-2105-7-123},
 journal = {BMC Bioinformatics},
 mendeley-groups = {Curve Fitting},
@@ -58,9 +59,10 @@ year = {2006}
 }
 
 @InProceedings{Liu2008,
-author={Liu, Yang and Wang, Wenping},
+author={Liu, Y. and Wang, W.},
 editor={Chen, Falai and J{\"u}ttler, Bert},
-title="A Revisit to Least Squares Orthogonal Distance Fitting of Parametric Curves and Surfaces",
+title={{A Revisit to Least Squares Orthogonal Distance Fitting of Parametric
+		Curves and Surfaces}},
 booktitle={Advances in Geometric Modeling and Processing},
 year={2008},
 publisher={Springer Berlin Heidelberg},
@@ -70,8 +72,9 @@ doi={10.1007/978-3-540-79246-8_29},
 isbn={978-3-540-79246-8}
 }
 @article{keles2018,
-title={Comparison of Classical Least Squares and Orthogonal Regression in Measurement Error Models.},
-author={Keles, Taliha},
+title={{Comparison of Classical Least Squares and Orthogonal Regression in
+		Measurement Error Models.}},
+author={Keles, T.},
 journal={International Online Journal of Educational Sciences},
 volume={10},
 number={3},
@@ -79,8 +82,8 @@ year={2018},
 url={http://www.iojes.net/Makaleler/386bbbe6-b98d-4766-a717-2c945a97f267.pdf}
 }
 @article{more1978levenberg,
-title={The Levenberg-Marquardt algorithm: Implementation and theory},
-author={Mor{\'e}, Jorge},
+title={{The Levenberg-Marquardt algorithm: Implementation and theory}},
+author={Mor{\'e}, J.},
 journal={Numerical Analysis},
 pages={105--116},
 year={1978},
