### 2019-05-24 dc0de0e
Fix argument in script.
