+++ test/scripts/draw.jl
@@ -24,15 +24,13 @@ n, model, modelstr = RAFF.model_list["logistic"]
 modl1 = (x) -> model(x, sol)
 modl2 = (x) -> model(x, θSol)
 
-t = minimum(x):0.01:maximum(x)
-PyPlot.plot(t, modl1.(t), "r", label="Ajustado RAFF")
+# t = minimum(x):0.01:maximum(x)
+# PyPlot.plot(t, modl1.(t), "r", label="Ajustado RAFF")
 
-θSol = [1000.0, 5000.0, 0.2, 3.0]
-PyPlot.plot(t, modl2.(t), "b--", label="Verdadeiro")
+# θSol = [1000.0, 5000.0, 0.2, 3.0]
+# PyPlot.plot(t, modl2.(t), "b--", label="Verdadeiro")
 
-PyPlot.legend(loc=4)
-
-c[9] = 1.0
+# PyPlot.legend(loc=4)
 
 PyPlot.scatter(x, y, c=c, marker="o", s=50.0, linewidths=0.2,
                cmap=PyPlot.cm."Paired", alpha=0.6)
