+++ joss/paper.bib
@@ -5,15 +5,15 @@ issn = {0925-5001},
 journal = {Journal of Global Optimization},
 number = {1},
 pages = {1--22},
-title = {{Low Order-Value Optimization and applications}},
+title = {{Low Order-Value Optimization and Applications}},
 url = {http://link.springer.com/10.1007/s10898-008-9280-3},
 volume = {43},
 year = {2009}
 }
 
 @article{Andreani2005,
-title={{Nonlinear-programming reformulation of the order-value optimization
-		problem}},
+title={{Nonlinear-Programming Reformulation of the Order-Value Optimization
+		Problem}},
 author={Andreani, R. and Dunder, C. and Mart{\'\i}nez, J.  M.},
 journal={Mathematical Methods of Operations Research},
 volume={61},
@@ -27,7 +27,7 @@ publisher={Springer}
 }
 
 @article{Martinez2012,
-title={{Generalized order-value optimization}},
+title={{Generalized Order-Value Optimization}},
 author={Mart{\'\i}nez, J. M.},
 journal={Top},
 volume={20},
@@ -53,7 +53,8 @@ doi = {10.1186/1471-2105-7-123},
 journal = {BMC Bioinformatics},
 mendeley-groups = {Curve Fitting},
 number = {123},
-title = {{Detecting outliers when fitting data with nonlinear regression – a new method based on robust nonlinear regression and the false discovery rate}},
+title = {{Detecting Outliers When Fitting Data with Nonlinear Regression – a New
+		Method Based on Robust Nonlinear Regression and the False Discovery Rate}},
 volume = {7},
 year = {2006}
 }
@@ -82,7 +83,7 @@ year={2018},
 url={http://www.iojes.net/Makaleler/386bbbe6-b98d-4766-a717-2c945a97f267.pdf}
 }
 @article{more1978levenberg,
-title={{The Levenberg-Marquardt algorithm: Implementation and theory}},
+title={{The Levenberg-Marquardt Algorithm: Implementation and Theory}},
 author={Mor{\'e}, J.},
 journal={Numerical Analysis},
 pages={105--116},
