### 2019-05-24 fsobral/RAFF.jl#18 @evcastelani
Hi @matbesancon! You're right. The title of the references had to be standardized. Thanks for the note. However, running locally by Pandoc, the use of `{{tilte}}` worked like a charm but in JOSS no. In this way, I made the manual correction of capital letters to avoid possible errors.
