### 2019-05-24 fsobral/RAFF.jl#18 @matbesancon
Hi, no it seems to have worked, the titles are properly capitalized. I think this was the last point on my side.
@abelsiqueira should we close this issue?
