### 2019-05-24 fsobral/RAFF.jl#18 @abelsiqueira
Yes, I'm happy with the changes.

On Fri, May 24, 2019, 10:41 Mathieu Besançon <notifications@github.com>
wrote:

> Hi, no it seems to have worked, the titles are properly capitalized. I
> think this was the last point on my side.
> @abelsiqueira <https://github.com/abelsiqueira> should we close this
> issue?
>
