### 2019-05-24 fsobral/RAFF.jl#18 @matbesancon
Great, doing one last check on the last version of the software in an hour or so, then validating
