### 2019-05-24 6223ec2
Add information about parameters and modify variable name to follow Julia sytle guide.
