### 2019-05-24 30cdbac
Add script for parallel run of RAFF.
