### 2019-05-24 2d0c283
Refactor drawing functions to draw found outliers and beautiful graphics.
