### 2019-05-24 4eed449
Remove tight stopping criterion.
