### 2019-05-28 fsobral/RAFF.jl#18 @matbesancon
sorry last small details in "installation and usage":
"and can be directly installed from *the* Julia package repository" or "and can be directly installed from Julia's package repository" but the former might be preferred.

"All *descriptions*" or "The whole description"
