### 2019-05-28 d6d9110
typos in installation and usage #18
