### 2019-05-28 fsobral/RAFF.jl#18 @evcastelani
@matbesancon you're right. These typos have been fixed.
