### 2019-05-28 fsobral/RAFF.jl#18 @matbesancon
ok great, the paper seems fine, watch out builds seem to all fail
