### 2019-05-28 fsobral/RAFF.jl#18 @evcastelani
@matbesancon Ok, thanks! Now we fixed the problem and both branches (master and joss) are passing the automated tests.
