### 2019-05-28 fsobral/RAFF.jl#18 @matbesancon
perfect, I think this is good on my side. @abelsiqueira do you still have something on the last version of the paper?
