### 2019-05-28 #1385 @matbesancon
@labarba @poulson my review is done, after the revisions made to the package and paper I recommend its acceptance for JOSS. It's a great package and the paper explains it well
