### 2019-05-29 fsobral/RAFF.jl#18 @abelsiqueira
Reference Keles needs to be updated (DOI, pages and author name):
```
@article{Kele2018,
  doi = {10.15345/iojes.2018.03.013},
  url = {https://doi.org/10.15345/iojes.2018.03.013},
  year = {2018},
  publisher = {International Online Journal of Educational Sciences},
  volume = {10},
  number = {3},
  author = {Taliha Kele{\c{s}}},
  title = {Comparison of Classical Least Squares and Orthogonal Regression in Measurement Error Models},
  journal = {International Online Journal of Educational Sciences}
}
```
Found the DOI on the paper and the bibtex using https://www.doi2bib.org/bib/10.15345/iojes.2018.03.013.

That's it for me.
