+++ joss/paper.bib
@@ -74,13 +74,14 @@ isbn={978-3-540-79246-8}
 }
 @article{keles2018,
 title={{Comparison of Classical Least Squares and Orthogonal Regression in
-		Measurement Error Models.}},
-author={Keles, T.},
+Measurement Error Models.}},
+author={Taliha Kele{\c{s}}},
 journal={International Online Journal of Educational Sciences},
 volume={10},
 number={3},
 year={2018},
-url={http://www.iojes.net/Makaleler/386bbbe6-b98d-4766-a717-2c945a97f267.pdf}
+pages={200--214},
+doi = {10.15345/iojes.2018.03.013}
 }
 @article{more1978levenberg,
 title={{The Levenberg-Marquardt Algorithm: Implementation and Theory}},
