### 2019-05-29 fsobral/RAFF.jl#18 @evcastelani
@abelsiqueira  you're right! We fixed DOI,name and pages as you suggest.
