### 2019-05-29 fsobral/RAFF.jl#18 @evcastelani
@matbesancon and @abelsiqueira  Can I close this issue?
