### 2019-05-29 fsobral/RAFF.jl#18 @matbesancon
If you re-built the paper with corrections from @abelsiqueira
<https://github.com/abelsiqueira> it's up to him to close, it's good on my
side
