### 2019-05-29 fsobral/RAFF.jl#18 @abelsiqueira
Yes, close it, thanks for the corrections.
