### 2019-05-29 #1385 @abelsiqueira
@labarba @poulson my review is also done, and I also recommend acceptance. The version number above, v0.4.1, has changed to v0.6.0. Otherwise, I'm good to go. Nice paper, @evcastelani and @fsobral.
