### 2019-05-29 #1385 @fsobral
Thanks @abelsiqueira and @matbesancon for the careful reading of the manuscript and for the suggestions to the software! It was greatly improved.

Regarding the version, I don't know how to change it...
