### 2019-05-30 #1385 @evcastelani
@abelsiqueira and @matbesancon, we appreciate the effort and time spent by you to help us improve this work. Sincerely, thank you very much for the corrections and suggestions.
