### 2019-06-05 d96213b
Add new parameter to function.
