### 2019-06-09 #1385 @poulson
:wave: @fsobral. Thank you for your patience for my response; my apologies for the delay!

I have a few comments on the manuscript:

* Given that the point of the package is learning a model for a black-box function, unqualified usage of the term "optimal" in the opening paragraphs seems a bit aggressive given that unregularized two-norm objective functions are not typically the most competitive in terms of generalization. It would be good to add some sort of qualifier there.

* In terms of demonstrating non-differentiability of the LOVO problem, it seems to me that one can easily show this with m=2 and p=1, given F1(theta) = theta and F2(theta) = -theta in the neighborhood of theta=0. Is there a need to spend half a page showing this?

* As a nit, I would recommend adding a comma at the position "is a continuous function but assuming that" to become "is a continuous function, but assuming that".

Also, would you mind creating a DOI for the submission on Zenodo and posting it here?
