### 2019-06-10 843b162
'optimal' term from the sentence 'optimal parameters' removed
