### 2019-06-10 #1385 @evcastelani
>  @fsobral. Thank you for your patience for my response; my apologies for the delay!
>
> I have a few comments on the manuscript:
>
> * Given that the point of the package is learning a model for a black-box function, unqualified usage of the term "optimal" in the opening paragraphs seems a bit aggressive given that unregularized two-norm objective functions are not typically the most competitive in terms of generalization. It would be good to add some sort of qualifier there.

We have removed the term from the phrase. Our use for "optimal" is in the optimization sense, i.e., the parameters come from the solution of an optimization problem. However, we are only able to achieve stationary points, so there is no point in calling them "optimal" if it may cause confusion.

> * In terms of demonstrating non-differentiability of the LOVO problem, it seems to me that one can easily show this with m=2 and p=1, given F1(theta) = theta and F2(theta) = -theta in the neighborhood of theta=0. Is there a need to spend half a page showing this?

You're right, your example is enough to show the non-differentiability, but our example is used to show the whole structure of LOVO problem in context of fitting functions, not just differentiability. Our example was motivated by the discussion in https://github.com/fsobral/RAFF.jl/issues/18.

> * As a nit, I would recommend adding a comma at the position "is a continuous function but assuming that" to become "is a continuous function, but assuming that".

Done.

>
> Also, would you mind creating a DOI for the submission on Zenodo and posting it here?

Ok, we are waiting for Zenodo's DOI for our github project.
