### 2019-06-10 #1385 @fsobral
@poulson We have successfully created a DOI for the package: 10.5281/zenodo.3242619
