### 2019-06-19 dc71433
Add uniform perturbation for random generator.
