### 2019-06-19 bb2787f
Add possibility to provide good initial guesses.
