### 2019-06-19 28a0dad
Count the number of problems where the outliers were correctly identified.
