+++ dev/search_index.js
@@ -141,7 +141,7 @@ var documenterSearchIndex = {"docs": [
     "page": "API",
     "title": "RAFF.raff",
     "category": "function",
-    "text": "raff(model::Function, data::Array{Float64, 2}, n::Int; MAXMS::Int=1,\n     SEEDMS::Int=123456789, initguess=zeros(Float64, n))\n\nraff(model::Function, gmodel!::Function, data::Array{Float64, 2}, n::Int;\n     [MAXMS::Int=1, SEEDMS::Int=123456789, initguess=zeros(Float64, n),\n      kwargs...])\n\nRobust Algebric Fitting Function (RAFF) algorithm. This function uses a voting system to automatically find the number of trusted data points to fit the model.\n\nmodel: function to fit data. Its signature should be given by\nmodel(x, θ)\nwhere x is the multidimensional argument and θ is the n-dimensional vector of parameters\ngmodel!: gradient of the model function. Its signature should be given by\ngmodel!(g, x, θ)\nwhere x is the multidimensional argument, θ is the n-dimensional vector of parameters and the gradient is written in g.\ndata: data to be fit. This matrix should be in the form\nx11 x12 ... x1N y1\nx21 x22 ... x2N y2\n:\nwhere N is the dimension of the argument of the model (i.e. dimension of x).\nn: dimension of the parameter vector in the model function\n\nThe optional arguments are\n\nMAXMS: number of multistart points to be used\nSEEDMS: integer seed for random multistart points\ninitialguess: a good guess for the starting point and for generating random points in the multistart strategy\nε: gradient stopping criteria to lmlovo\nnoutliers: integer describing the maximum expected number of outliers. The default is half.\n\nReturns a RAFFOutput object with the best parameter found.\n\n\n\n\n\n"
+    "text": "raff(model::Function, data::Array{Float64, 2}, n::Int; MAXMS::Int=1,\n     SEEDMS::Int=123456789, initguess::Vector{Float64}=zeros(Float64, n))\n\nraff(model::Function, gmodel!::Function, data::Array{Float64, 2}, n::Int;\n     [MAXMS::Int=1, SEEDMS::Int=123456789,\n     initguess::Vector{Float64}=zeros(Float64, n), ε::Float64=1.0e-4,\n     noutliers::Int=-1])\n\nRobust Algebric Fitting Function (RAFF) algorithm. This function uses a voting system to automatically find the number of trusted data points to fit the model.\n\nmodel: function to fit data. Its signature should be given by\nmodel(x, θ)\nwhere x is the multidimensional argument and θ is the n-dimensional vector of parameters\ngmodel!: gradient of the model function. Its signature should be given by\ngmodel!(g, x, θ)\nwhere x is the multidimensional argument, θ is the n-dimensional vector of parameters and the gradient is written in g.\ndata: data to be fit. This matrix should be in the form\nx11 x12 ... x1N y1\nx21 x22 ... x2N y2\n:\nwhere N is the dimension of the argument of the model (i.e. dimension of x).\nn: dimension of the parameter vector in the model function\n\nThe optional arguments are\n\nMAXMS: number of multistart points to be used\nSEEDMS: integer seed for random multistart points\ninitialguess: a good guess for the starting point and for generating random points in the multistart strategy\nε: gradient stopping criteria to lmlovo\nnoutliers: integer describing the maximum expected number of outliers. The default is half.\n\nReturns a RAFFOutput object with the best parameter found.\n\n\n\n\n\n"
 },
 
 {
@@ -149,7 +149,7 @@ var documenterSearchIndex = {"docs": [
     "page": "API",
     "title": "RAFF.praff",
     "category": "function",
[…]
