### 2019-06-26 c6324a1
Change arguments of scatter, to avoid warning.
