### 2019-06-26 b634a7f
Add choice for model
