### 2019-06-26 8b64888
Fix problem of sending solution vector as argument.
