### 2019-06-26 063c6d5
Print solution in debug mode, for each value of p
