### 2019-06-26 9c2afe7
Replace 2 norm by Inf norm in voting system.
