### 2019-06-26 59ce3ca
Add information about the average number of detected outliers.
