### 2019-06-26 17fafa9
Improve colors in plots and position of the legend.
