+++ test/scripts/draw.jl
@@ -2,9 +2,28 @@ using DelimitedFiles
 using PyPlot
 using RAFF
 
-function draw_problem(;raff_output=nothing, model_str="logistic")
+"""
 
-    datafile = "/tmp/output.txt"
+    draw_problem(;raff_output=nothing, model_str="logistic",
+                  datafile="/tmp/output.txt")
+
+Draw the problem data. By default it is assumed that the model is
+given by the logistic model and the data file is given in file
+`/tmp/output.txt`.
+
+If a [RAFFOutput](@ref) object is provided, then it plots the model
+found and also the outliers (true and false positives).
+
+Optional arguments:
+
+  - `raff_output`: [RAFFOutput](@ref) object with the solution
+    obtained
+  - `model_str`: a string with the name of the model to be used to
+    plot the solution. See [model_list](@ref) for details.
+  - `datafile`: file containing the data.
+
+"""
+function draw_problem(;raff_output=nothing, model_str="logistic", datafile="/tmp/output.txt")
 
     fp = open(datafile, "r")
 
@@ -16,14 +35,14 @@ function draw_problem(;raff_output=nothing, model_str="logistic")
 
     x = M[:, 1]
     y = M[:, 2]
-    c = M[:, 3]
+    co = M[:, 3]
 
-    true_outliers = findall(c .== 1)
+    true_outliers = findall(co .== 1)
 
-    PyPlot.scatter(x[c .== 0.0], y[c .== 0.0], color=PyPlot.cm."Pastel1"(1.0),
+    PyPlot.scatter(x[co .== 0.0], y[co .== 0.0], color=PyPlot.cm."Pastel1"(2.0/9.0),
                    marker="o", s=50.0, linewidths=0.2)
 
-    PyPlot.scatter(x[c .== 1.0], y[c .== 1.0], color=PyPlot.cm."Pastel1"(1.0),
+    PyPlot.scatter(x[co .== 1.0], y[co .== 1.0], color=PyPlot.cm."Pastel1"(2.0/9.0),
                    marker="^", s=50.0, linewidths=0.2, label="Outliers")
 
     if raff_output != nothing
@@ -39,21 +58,21 @@ function draw_problem(;raff_output=nothing, model_str="logistic")
         
         true_positives = intersect(true_outliers, raff_output.outliers)
         false_positives = setdiff(raff_output.outliers, true_positives)
-        
+
         PyPlot.scatter(x[false_positives], y[false_positives],
-                   color=PyPlot.cm."Pastel1"(0.0/9.0), marker="o", s=50.0,
-                   linewidths=0.2, label="False positives")
+                       color=PyPlot.cm."Pastel1"(0.0/9.0), marker="o",
+                       linewidths=0.2, edgecolors="k", s=50.0, label="False positives")
         
         PyPlot.scatter(x[true_positives], y[true_positives],
                        color=PyPlot.cm."Pastel1"(0.0/9.0), marker="^",
-                       s=50.0, linewidths=0.2, label="Identified outliers")
+                       s=50.0, linewidths=0.2, edgecolors="k", label="Identified outliers")
 
     end
 
     # θSol = [1000.0, 5000.0, 0.2, 3.0]
     # PyPlot.plot(t, modl2.(t), "b--", label="Verdadeiro")
 
-    PyPlot.legend(loc=4)
+    PyPlot.legend(loc="best")
 
     PyPlot.show()
     
