### 2019-06-28 #1385 @evcastelani
Dear @poulson, Is there any additional information, clarification on the text or on the current package that we can provide to help the evaluation of the submitted paper? In addition, as highlighted by @fsobral we don't know how change the version of package in JOSS system, can you help us?
