### 2019-06-28 a1208ea
Add tests for generator functions. Create memory efficient functions.
