### 2019-06-28 2a7e230
Create function to generate clustered noisy data.
