### 2019-06-29 297173d
Replaced Union by AbstractArray to deal with SubArrays.
