### 2019-06-29 afc6fe4
Avoid repetition of the values in the domain.
