### 2019-06-30 8310f3f
Add all optional arguments of generate_noisy_data to the clustered case.
