### 2019-06-30 5904ba7
Add new test to address the problem of not generating the full interval.
