### 2019-07-01 #1385 @poulson
(pinging @openjournals/joss-eics instead of Arfon)
