### 2019-07-01 #1385 @kyleniemeyer
Hi @fsobral and @evcastelani, I just made some minor edits to the paper in https://github.com/fsobral/RAFF.jl/pull/20, could you merge that? Also, there are a few final things:

- One of the expectations for JOSS papers is to contain
> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
Can you add this?
- Can you merge the `joss` branch into `master`, and then update the Zenodo archive with the updated version? Also, the metadata on the Zenodo archive needs to be cleaned up a bit, as the author names are not formatted properly (they should match the paper).

Thanks!
