### 2019-07-01 #1385 @evcastelani
Dear @kyleniemeyer

Thank you for your corrections in the paper.

> * One of the expectations for JOSS papers is to contain
>
> > A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
> > Can you add this?

Done

> * Can you merge the `joss` branch into `master`, and then update the Zenodo archive with the updated version? Also, the metadata on the Zenodo archive needs to be cleaned up a bit, as the author names are not formatted properly (they should match the paper).

We have merged the `joss` branch into `master`. Since our project is a registered Julia package, we are waiting for it to be added. When the process finishes, a new release will be automatically generated and the DOI will also be automatically generated. We will ping you with the new DOI as soon as it is generated.
