### 2019-07-01 f30f908
Added new method to generate random clustered problems.
