### 2019-07-01 93c9b9a
Replace xMin and xMax variables by a tuple. Improve comments.
