+++ src/generator.jl
@@ -30,9 +30,17 @@ const model_list = Dict(
 """
 
     generate_test_problems(datFilename::String, solFilename::String,
-                         model::Function, modelStr::String, n::Int,
-                         np::Int, p::Int; xMin=-10.0, xMax=10.0,
-                         θSol=10.0 * randn(n), std=200.0, out_times=7.0)
+        model::Function, modelStr::String, n::Int, np::Int, p::Int;
+        x_interval::Tuple{Float64, Float64}=(-10.0, 10.0),
+        θSol::Vector{Float64}=10.0 * randn(n), std::Float64=200.0,
+        out_times::Float64=7.0)
+
+    generate_test_problems(datFilename::String, solFilename::String,
+        model::Function, modelStr::String, n::Int, np::Int, p::Int,
+        cluster_interval::Tuple{Float64, Float64};
+        x_interval::Tuple{Float64, Float64}=(-10.0, 10.0),
+        θSol::Vector{Float64}=10.0 * randn(n), std::Float64=200.0,
+        out_times::Float64=7.0)
 
 Generate random data files for testing fitting problems.
 
@@ -51,21 +59,23 @@ Generate random data files for testing fitting problems.
   - `p` is the number of trusted points to be used in the LOVO
     approach.
 
+If `cluster_interval` is provided, then generates outliers only in
+this interval.
+
 Additional parameters:
 
-  - `xMin`, `xMax`: minimum and maximum interval for generating points
-    in one dimensional tests
+  - `x_interval`: interval for generating points in one dimensional
+    tests
   - `θSol`: true solution, used for generating perturbed points
   - `std`: standard deviation
   - `out_times`: deviation for outliers will be `out_times * std`.
 
 """
 function generate_test_problems(datFilename::String,
-                              solFilename::String, model::Function,
-                              modelStr::String,
-                              n::Int, np::Int, p::Int;
-                              xMin=-10.0, xMax=10.0,
-                              θSol=10.0 * randn(n), std=200.0, out_times=7.0)
+    solFilename::String, model::Function, modelStr::String, n::Int,
+    np::Int, p::Int; x_interval::Tuple{Float64, Float64}=(-10.0,
+    10.0), θSol::Vector{Float64}=10.0 * randn(n), std::Float64=200.0,
+    out_times::Float64=7.0)
 
     # Generate solution file
     
@@ -84,7 +94,7 @@ function generate_test_problems(datFilename::String,
     open(datFilename, "w") do data
     
         vdata, θsol, outliers = generate_noisy_data(model, n, np, p;
-                                   xMin=xMin, xMax=xMax, θSol=θSol,
+                                   x_interval=x_interval, θSol=θSol,
                                    std=std, out_times=out_times)
     
         # Dimension of the domain of the function to fit
@@ -104,8 +114,9 @@ end
 function generate_test_problems(datFilename::String,
     solFilename::String, model::Function, modelStr::String, n::Int,
     np::Int, p::Int, cluster_interval::Tuple{Float64, Float64};
-    xMin=-10.0, xMax=10.0, θSol=10.0 * randn(n), std=200.0,
-    out_times=7.0)
+    x_interval::Tuple{Float64, Float64}=(-10.0, 10.0),
+    θSol::Vector{Float64}=10.0 * randn(n), std::Float64=200.0,
+    out_times::Float64=7.0)
 
     # Generate solution file
     
[…]
