### 2019-07-01 09bb5c8
Refactor generator.jl function arguments.
