### 2019-07-01 46b6913
Add new function for drawing problems.
