### 2019-07-01 43d3a95
Add new option to script and fix error in clustered generator.
