+++ test/scripts/generate_fit_tests.jl
@@ -1,6 +1,7 @@
 using RAFF
 using Printf
 using ArgParse
+using Random
 
 function mainf()
 
@@ -32,6 +33,11 @@ function mainf()
         help = "Solution"
         nargs = '*'
         arg_type = Float64
+
+        "--cint"
+        help = "Cluster interval"
+        nargs = 2
+        arg_type = Float64
         
     end
 
@@ -41,15 +47,13 @@ function mainf()
 
     n, model, modelStr = RAFF.model_list[parsed_args["model"]]
 
-    xMin = - 0.0
-
-    xMax =   30.0
+    x_interval = (- 0.0, 30.0)
 
     θSol = Vector{Float64}(undef, n)
 
     if length(parsed_args["sol"]) == 0
 
-        map!(x->randn(), θSol, θSol)
+        randn!(θSol)
 
     elseif length(parsed_args["sol"]) == n
 
@@ -67,11 +71,23 @@ function mainf()
 
         ffname = "/tmp/" * fname * ".txt"
         fsname = "/tmp/" * fname * "_sol.txt"
-        
-        generate_test_problems(ffname, fsname, model, modelStr, n,
-            parsed_args["np"], parsed_args["p"]; x_interval=(xMin,
-            xMax), θSol=θSol, std=200.0, out_times=10.0)
 
+        if length(parsed_args["cint"]) == 2
+
+            generate_test_problems(ffname, fsname, model, modelStr, n,
+                parsed_args["np"], parsed_args["p"], x_interval,
+                Tuple(parsed_args["cint"]); θSol=θSol, std=200.0,
+                out_times=10.0)
+
+        else
+            
+            generate_test_problems(ffname, fsname, model, modelStr, n,
+                parsed_args["np"], parsed_args["p"];
+                x_interval=x_interval, θSol=θSol, std=200.0,
+                out_times=10.0)
+
+        end
+ 
         @printf("Created problem and solution files.\n")
 
     catch e
