### 2019-07-01 bc111f1
Add function to start massive tests.
