### 2019-07-02 364c7bb
updated references in paper.bib
