+++ joss/paper.bib
@@ -27,23 +27,22 @@ publisher={Springer}
 }
 
 @article{Martinez2012,
-title={{Generalized Order-Value Optimization}},
+title={Generalized Order-Value Optimization},
 author={Mart{\'\i}nez, J. M.},
 journal={Top},
 volume={20},
 number={1},
 pages={75--98},
 year={2012},
-doi = {10.1007/s1175001001691},
+doi = {10.1007/s11750-010-0169-1},
 publisher={Springer}
 }
 
-@article{Revels2016,
-    title = {{Forward-Mode Automatic Differentiation in Julia}},
+@misc{Revels2016,
+    title = {Forward-Mode Automatic Differentiation in {Julia}},
    author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
-  journal = {arXiv:1607.07892 [cs.MS]},
-     year = {2016},
-     url = {https://arxiv.org/abs/1607.07892}
+  howpublished = {\tt \href{https://arxiv.org/abs/1607.07892}{arXiv:1607.07892} [cs.MS]},
+     year = 2016,
 }
 
 
