### 2019-07-02 #1385 @kyleniemeyer
Hi @fsobral, there was an incorrect DOI, that I've corrected in https://github.com/fsobral/RAFF.jl/pull/21
