### 2019-07-02 #1385 @kyleniemeyer
Congrats to @fsobral and @evcastelani on your submission's publication in JOSS! And thanks very much to @abelsiqueira and @matbesancon for reviewing, and @poulson for editing.
