### 2019-07-02 #1385 @evcastelani
@kyleniemeyer thank you for accepting our article and providing important improvements. It was the first time we published on JOSS and it was a great experience. Thanks again to the reviewers @matbesancon
 and @abelsiqueira for all suggestions and the editor @poulson for the support.
