### 2019-07-10 #1385 @fsobral
Dear @kyleniemeyer , we forgot to mention the funding agency of our project in the paper. This is very important for us, since we might have to refund them or even be unable to apply again for funding. Is it possible to add such information and edit the paper? I'm really sorry for this trouble.
