### 2019-07-10 #1385 @fsobral
Thank you @arfon and @kyleniemeyer for the quick answer. The file `paper.md` has been updated in the master branch.
