### 2019-07-11 #1385 @arfon
OK, the paper is now updated but it might take a while for your browser to pick up the updated PDF as there's caching on the JOSS site.
